requirements.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
numpy==1.18.1
opencv-python==3.4.0.12
Pillow==7.1.2
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
